fix(arena): complete CodeRabbit follow-ups for Req 18 snapshot
- Extract shared isomorphic endpoint policy (arenaEndpointPolicy) used by
  snapshot helpers and ssrfGuard IPv4 ranges; always reject metadata/.local
- Make snapshot helpers browser-safe (no process.env reads); loopback only via opts
- Harden Olive list concurrent download with ref guard
- React 19 Button ref-as-prop (drop forwardRef)
- Catalog sync test for DEFAULT_BASE_URLS vs registerProvider defaults
- Stronger 403 message assertion; Property 21 modelId trim
