+++ src/lib/arenaAssistantSnapshot.ts
@@ -1,20 +1,32 @@
 /**
  * Arena Assistant cloud-snapshot helpers (Req 18).
  *
- * Pure client/server-shareable gates + mappers. Outbound URL policy mirrors
- * `assertUrlPolicy` in ssrfGuard (protocol / credentials / private literals /
- * loopback-http override) without Node-only DNS pinning.
+ * Pure client/server-shareable gates + mappers. Outbound URL policy lives in
+ * `arenaEndpointPolicy.ts` (shared with server ssrfGuard for IPv4 ranges).
+ *
+ * Browser-safe: this module never reads `process.env`. Callers that need the
+ * loopback-HTTP override (server route only) pass `{ allowLoopbackHttp: true }`.
  */
 
 import type { ArenaSlotConfig } from "@/lib/stores/playgroundStore";
+import {
+  assertArenaEndpointUrlPolicy,
+  type ArenaEndpointPolicyOpts,
+} from "@/lib/arenaEndpointPolicy";
 
 /** Providers that are not OpenAI chat-completions shaped for Arena cloud runs. */
 // copilot: Arena cloud-inference only sends Authorization + Content-Type; Copilot
 // needs provider-specific headers, so one-click snapshot cannot work.
 const NON_OPENAI_COMPAT = new Set(["gemini", "anthropic", "devin", "codex", "copilot"]);
 
-/** Catalog defaults when the active config omits baseUrl. */
-const DEFAULT_BASE_URLS: Record<string, string> = {
+/**
+ * Catalog defaults when the active config omits baseUrl.
+ * Server snapshot prefers `plugin.defaultBaseUrl` first; this table is the
+ * pure-helper fallback used by unit tests and offline resolution.
+ * Kept in sync with `src/server/services/ai/openai.ts` registerProvider defaults
+ * via `arenaAssistantSnapshot.catalog.test.ts`.
+ */
+export const DEFAULT_BASE_URLS: Record<string, string> = {
   openai: "https://api.openai.com/v1",
   mistral: "https://api.mistral.ai/v1",
   xai: "https://api.x.ai/v1",
@@ -61,162 +73,23 @@ function stripTrailingSlashes(url: string): string {
   return url.replace(/\/+$/, "");
 }
 
-function stripBrackets(hostname: string): string {
-  return hostname.replace(/^\[|\]$/g, "").toLowerCase();
-}
-
-function isLoopbackHostname(hostname: string): boolean {
-  const host = stripBrackets(hostname);
-  return host === "localhost" || host === "127.0.0.1" || host === "::1";
-}
-
-function isBlockedIpv4(ip: string): boolean {
-  const m = /^(\d{1,3})\.(\d{1,3})\.(\d{1,3})\.(\d{1,3})$/.exec(ip);
-  if (!m) return false;
-  const [a, b, c] = [Number(m[1]), Number(m[2]), Number(m[3])];
-  if ([a, b, c, Number(m[4])].some((n) => n > 255)) return true;
-  if (a === 0) return true;
-  if (a === 10) return true;
-  if (a === 127) return true;
-  if (a === 169 && b === 254) return true;
-  if (a === 172 && b >= 16 && b <= 31) return true;
-  if (a === 192 && b === 168) return true;
-  if (a === 100 && b >= 64 && b <= 127) return true;
-  if (a === 192 && b === 0 && c === 0) return true;
-  if (a === 192 && b === 0 && c === 2) return true;
-  if (a === 198 && (b === 18 || b === 19)) return true;
-  if (a === 198 && b === 51 && c === 100) return true;
-  if (a === 203 && b === 0 && c === 113) return true;
-  if (a >= 224) return true;
-  return false;
-}
-
-function isLoopbackIp(ip: string): boolean {
-  const host = stripBrackets(ip);
-  return host === "127.0.0.1" || host === "::1" || host.startsWith("127.");
-}
-
-function parseIpv6(host: string): bigint | null {
-  if (!host.includes(":")) return null;
-  const parts = host.split("::");
-  if (parts.length > 2) return null;
-  const expand = (value: string): string[] => {
-    if (!value) return [];
-    const items = value.split(":");
-    const last = items[items.length - 1];
-    if (last.includes(".")) {
-      const octets = last.split(".").map(Number);
-      if (octets.length !== 4 || octets.some((n) => !Number.isInteger(n) || n > 255)) return [];
-      items.splice(
-        items.length - 1,
-        1,
-        ((octets[0] << 8) | octets[1]).toString(16),
-        ((octets[2] << 8) | octets[3]).toString(16),
-      );
-    }
-    return items;
-  };
-  const left = expand(parts[0]);
-  const right = expand(parts[1] ?? "");
-  const zeros = 8 - left.length - right.length;
-  if (
-    (parts.length === 1 && zeros !== 0) ||
-    zeros < 0 ||
-    [...left, ...right].some((p) => !/^[0-9a-f]{1,4}$/i.test(p))
-  ) {
-    return null;
-  }
-  // Zero-pad each hextet so the 128-bit value is always 32 hex digits.
-  // Unpadded joins (e.g. fe80::1 → "fe800…001") shift high bits and break
-  // link-local / ULA / mapped checks.
-  const groups = [...left, ...Array(zeros).fill("0"), ...right].map((g) =>
-    g.padStart(4, "0").toLowerCase(),
-  );
-  return BigInt(`0x${groups.join("")}`);
-}
-
-function isBlockedIpv6(host: string): boolean {
-  const value = parseIpv6(host);
-  if (value === null) return false;
-  const first = Number(value >> 120n);
-  const second = Number((value >> 118n) & 0x3n);
-  // loopback, link-local, ULA, and IPv4-mapped/compatible addresses
-  return value === 1n || (first === 0xfe && second === 0x2) || first >= 0xfc && first <= 0xfd || value >> 32n === 0n || value >> 32n === 0xffffn;
-}
-
-function isLoopbackIpv6(host: string): boolean {
-  const value = parseIpv6(host);
-  if (value === 1n) return true;
-  if (value === null || value >> 32n !== 0n) return false;
-  const ipv4 = Number(value & 0xffffffffn);
-  return isLoopbackIp(`${ipv4 >>> 24}.${(ipv4 >>> 16) & 255}.${(ipv4 >>> 8) & 255}.${ipv4 & 255}`);
-}
-
-/**
- * Shape-level outbound policy aligned with Arena `assertUrlPolicy`
- * (no DNS resolve — literal hosts only).
- */
-export function assertArenaEndpointUrlPolicy(
-  rawUrl: string,
-  opts?: { allowLoopbackHttp?: boolean },
-): void {
-  const allowLoopbackHttp = opts?.allowLoopbackHttp ?? false;
-  let url: URL;
-  try {
-    url = new URL(rawUrl);
-  } catch {
-    throw new Error("Invalid endpointUrl");
-  }
-  if (url.protocol !== "http:" && url.protocol !== "https:") {
-    throw new Error("Only http/https endpoints are supported");
-  }
-  if (url.username || url.password) {
-    throw new Error("Credentialed endpoints are not supported");
-  }
-  const host = stripBrackets(url.hostname);
-  if (!host) throw new Error("Invalid endpointUrl");
-
-  if (host === "metadata.google.internal" || host.endsWith(".local")) {
-    if (!(allowLoopbackHttp && isLoopbackHostname(host))) {
-      throw new Error("Private endpoints are not supported");
-    }
-  }
-
-  const loopbackName = isLoopbackHostname(host);
-  const allowLoopback = allowLoopbackHttp && loopbackName && url.protocol === "http:";
-
-  if (url.protocol !== "https:" && !allowLoopback) {
-    throw new Error("HTTPS endpoints are required");
-  }
-
-  // Literal IPv4 and IPv6 (including IPv4-mapped/compatible forms).
-  if (/^\d{1,3}(?:\.\d{1,3}){3}$/.test(host)) {
-    if (isBlockedIpv4(host) && !(allowLoopback && isLoopbackIp(host))) {
-      throw new Error("Private endpoints are not supported");
-    }
-  } else if (isBlockedIpv6(host) && !(allowLoopback && isLoopbackIpv6(host))) {
-    throw new Error("Private endpoints are not supported");
-  }
-}
-
 /**
  * True when the provider can feed Arena cloud inference (OpenAI chat shape +
  * outbound URL policy). Does not require a non-empty apiKey — local engines
  * may omit keys.
+ *
+ * Never reads `process.env` (browser-safe). Pass `allowLoopbackHttp` from the
+ * server when `OLIVE_ALLOW_LOOPBACK_HTTP` is set.
  */
 export function isArenaOpenAiCompatProvider(
   provider: {
     provider: string;
     baseUrl?: string | null;
   },
-  opts?: { allowLoopbackHttp?: boolean },
+  opts?: ArenaEndpointPolicyOpts,
 ): boolean {
   if (!provider.provider || NON_OPENAI_COMPAT.has(provider.provider)) return false;
-  const endpoint = resolveArenaSnapshotEndpointUrl(provider, {
-    allowLoopbackHttp:
-      opts?.allowLoopbackHttp ?? process.env.OLIVE_ALLOW_LOOPBACK_HTTP === "true",
-  });
-  return endpoint !== null;
+  return resolveArenaSnapshotEndpointUrl(provider, opts) !== null;
 }
 
 /**
@@ -224,7 +97,7 @@ export function isArenaOpenAiCompatProvider(
  */
 export function resolveArenaSnapshotEndpointUrl(
   provider: { provider: string; baseUrl?: string | null },
-  opts?: { allowLoopbackHttp?: boolean },
+  opts?: ArenaEndpointPolicyOpts,
 ): string | null {
   if (!provider.provider || NON_OPENAI_COMPAT.has(provider.provider)) return null;
 
@@ -248,7 +121,7 @@ export function resolveArenaSnapshotEndpointUrl(
  */
 export function buildAssistantCloudSnapshot(
   provider: ArenaProviderDescriptor | null | undefined,
-  opts?: { allowLoopbackHttp?: boolean },
+  opts?: ArenaEndpointPolicyOpts,
 ): AssistantCloudSnapshotResponse {
   if (!provider?.provider) {
     return { eligible: false, reason: "No Assistant provider configured" };
